requirements.txt
requests
python-dotenv
groq

